# lib/api-zod/src/generated/types/inventoryItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryItem {
  ID: number;
  /** @nullable */
  BoxNo?: string | null;
  /** @nullable */
  BoxTypeId?: string | null;
  /** @nullable */
  SkuCode?: string | null;
  /** @nullable */
  ItemName?: string | null;
  /** @nullable */
  Qty?: number | null;
  /** @nullable */
  BatchNoSrNo?: string | null;
  /** @nullable */
  MfgDate?: string | null;
  /** @nullable */
  ExpDate?: string | null;
  /** @nullable */
  CompanyName?: string | null;
}