# lib/api-zod/src/generated/types/item.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Item {
  id: number;
  /** @nullable */
  sNo?: string | null;
  /** @nullable */
  cubeID?: string | null;
  /** @nullable */
  cubeName?: string | null;
  /** @nullable */
  frameID?: string | null;
  /** @nullable */
  frameName?: string | null;
  /** @nullable */
  boxID?: string | null;
  /** @nullable */
  boxName?: string | null;
  /** @nullable */
  kitID?: string | null;
  /** @nullable */
  kitName?: string | null;
  /** @nullable */
  kitQty?: string | null;
  /** @nullable */
  itemID?: string | null;
  /** @nullable */
  itemName?: string | null;
  /** @nullable */
  itemQty?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  category?: string | null;
}