# lib/api-zod/src/generated/types/kit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Kit {
  kitID: string;
  /** @nullable */
  kitName?: string | null;
  /** @nullable */
  kitQty?: string | null;
  /** @nullable */
  boxName?: string | null;
  /** @nullable */
  frameName?: string | null;
  /** @nullable */
  cubeName?: string | null;
  itemCount: number;
}